Crate Nova

Nova

The Nova kernel is a component of CharlotteOS, an experimental modern operating system. The kernel is responsible for initializing the hardware, providing commonizing abstractions for all hardware resources, and managing the execution of user-space applications and the environment in which they run. It is a crucial part of the operating system, as it provides the foundation on which the rest of the system is built and it touches every hardware and software component of the system on which it is used.

Modules

common

Common Utilities for use throughout Nova

console

The Kernel Console

environment

hal

Hardware Abstraction Layer

init

log

Kernel Logging Macros

memory

Macros

log

logln

Functions

main⚠

This is the entry point for the kernel. The main function is called by the bootloader after setting up the environment. It is made C ABI compatible so that it can be called by Limine or any other Limine Boot Protocol compliant bootloader.

panic 🔒